requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dbus-python==1.3.2
evdev==1.6.1
idna==3.4
joblib==1.2.0
nltk==3.8.1
notify2==0.3.1
playsound==1.3.0
pycairo==1.24.0
PyGObject==3.44.1
pynput==1.7.6
python-dotenv==1.0.0
python-xlib==0.33
regex==2023.6.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.3
